requirements.txt
Flask==2.2.3
openai==0.26.5
Flask-Cors==3.0.10
Werkzeug==2.2.2
python-dotenv==1.0.0